casperjs tests for wrappers
Summary:
This is a draft or rather a proof of concept on testing the razorflow server-side bindings. This includes a simple implementation that does some basic integration tests on the wrappers.

initial implementation of casper wrapper tester

show error messages if baseurl and testspath are not specified

added assertSelectorTextEquals method

Test Plan: no test plan

Reviewers: anirudh, nari, Viral

Differential Revision: http://phabricator.razorflow.com/D5